change eslint code style to accommodate file extension. refractor gulpfile and made little changes in the index file
