+++ index.html
@@ -45,7 +45,7 @@
     </div>
   </footer>
 
-  <script src="app/js/weather.js"></script>
+  <script type="module" src="/app/js/weather.js" async></script>
   <script src="https://kit.fontawesome.com/8c64d9042a.js" crossorigin="anonymous"></script>
 </body>
 </html>
